+++ CHANGELOG.md
@@ -19,6 +19,7 @@ _All Hearthstones in a player's inventory, and ender chest, share the same coold
 ### Changed
 
 - Hearthstones now teleport the player to a safe position in front of the bound Fireplace.
+- Changed item groups for Hearthstones and Fireplaces to a custom Inns & Innkeepers group.
 
 ### Fixed